pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bamboo-shoot-soup-backend"
version = "0.1.0"
description = "MVP backend scaffold for Bamboo Shoot Soup."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "sqlalchemy>=2.0.38",
  "alembic>=1.16.0",
  "pydantic-settings>=2.7.0",
  "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "httpx>=0.28.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
markers = [
  "live_llm: tests that call real LLM endpoints and require RUN_LIVE_LLM_TESTS=1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
